+++ R/data.R
@@ -1,4 +1,4 @@
-#' Imprimerie datalist
+#' Example datalist
 #'
 #' Fictional dataset containing the necessary information to run an equal pay
 #' analysis.
@@ -11,7 +11,7 @@
 #'   \item{years_of_service}{years of service, in years},
 #'   \item{training}{training code, 1-8},
 #'   \item{professional_function}{function / job},
-#'   \item{skill_level}{skill level code, 1-4},
+#'   \item{level_of_requirements}{level of requirements code, 1-4},
 #'   \item{professional_position}{professional position / hierarchy code, 1-5},
 #'   \item{activity_rate}{activity rate, in percent},
 #'   \item{paid_hours}{paid hours, in hours},
@@ -28,7 +28,7 @@
 #'   \item{supplement3}{additional remarks (3 of 5)},
 #'   \item{supplement4}{additional remarks (4 of 5)},
 #'   \item{supplement5}{additional remarks (5 of 5)}}
-"datalist_imprimerie"
+"datalist_example"
 
 
 #' Column names