+++ R/data_check.R
@@ -20,9 +20,10 @@ check_data <- function(data) {
   idx <- rownames(data)
 
   # ----- Missing values check -------------------------------------------------
+
   # Check for missing values of all obligatory columns
   obligatory_columns <- c("personal_number", "age", "sex", "years_of_service",
-                          "training", "professional_function", "skill_level",
+                          "training", "professional_function", "level_of_requirements",
                           "professional_position", "basic_wage", "population")
   for (col in obligatory_columns) {
     error_rows <- as.numeric(idx[is.na(data[, col])])
@@ -34,6 +35,7 @@ check_data <- function(data) {
   }
 
   # ----- Correct values check -------------------------------------------------
+
   # Check for non-unique personal_number
   error_rows <- as.numeric(idx[duplicated(data$personal_number) |
                       duplicated(data$personal_number, fromLast = TRUE)])
@@ -49,18 +51,21 @@ check_data <- function(data) {
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$age[error_rows], "age",
                                "'Age' is not a whole number", 1))
+
   # Check for age limits (>= 13, <= 100)
   error_rows <- as.numeric(idx[data$age < 13 | data$age > 100])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$age[error_rows], "age",
                                "'Age' is not between 13 and 100", 1))
+
   # Check for plausible age limits (>= 15, <= 70)
   error_rows <- as.numeric(idx[data$age < 15 | data$age > 70])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$age[error_rows], "age",
                                "'Age' is not between 15 and 70", 2))
+
   # Check for incorrect sex (neither F nor M)
   error_rows <- as.numeric(idx[!(data$sex %in% c("F", "M"))])
   errors <- rbind(errors,
@@ -76,6 +81,7 @@ check_data <- function(data) {
                                data$years_of_service[error_rows],
                                "years_of_service",
                                "'Years of service' is not between 0 and 85", 1))
+
   # Check for plausible years of service limits (>= 0, <= 55)
   error_rows <- as.numeric(idx[data$years_of_service > 55 &
                                  data$years_of_service <= 85])
@@ -84,6 +90,22 @@ check_data <- function(data) {
                                data$years_of_service[error_rows],
                                "years_of_service",
                                "'Years of service' is more than 55", 2))
+
+  # Check for entry age
+  error_rows <- as.numeric(idx[data$age - data$years_of_service < 13])
+  errors <- rbind(errors,
+                  build_errors(error_rows, data$personal_number[error_rows],
+                               data$age[error_rows] - data$years_of_service[error_rows], "entry_age",
+                               "Entry age is less than 13", 1))
+
+  # Check for implausible entry age
+  error_rows <- as.numeric(idx[data$age - data$years_of_service < 16 &
+                                 data$age - data$years_of_service >= 13])
+  errors <- rbind(errors,
+                  build_errors(error_rows, data$personal_number[error_rows],
+                               data$age[error_rows] - data$years_of_service[error_rows], "entry_age",
+                               "Entry age is less than 16", 2))
+
   # Check for wrong training/education values
   error_rows <- as.numeric(idx[!(data$training %in% 1:8)])
   errors <- rbind(errors,
@@ -91,27 +113,36 @@ check_data <- function(data) {
                                data$training[error_rows],
                                "training", "'Training' is not between 1 and 8",
                                1))
+
   # Check for FTE limits (0 - 150% or 0 - 300 hours)
-  error_rows <- idx[data$activity_rate > 150 | data$activity_rate < 0]
+  error_rows <- as.numeric(idx[data$activity_rate > 150 | data$activity_rate < 0])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$activity_rate[error_rows], "activity_rate",
                                "'Activity rate' is not between 0% and 150%", 1))
-  error_rows <- idx[data$activity_rate > 100 & data$activity_rate <= 150]
+  error_rows <- as.numeric(idx[data$activity_rate > 100 & data$activity_rate <= 150])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$activity_rate[error_rows], "activity_rate",
                                "'Activity rate' is above 100%", 2))
-  error_rows <- idx[data$paid_hours > 300 | data$paid_hours < 0]
+  error_rows <- as.numeric(idx[data$paid_hours > 300 | data$paid_hours < 0])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$paid_hours[error_rows], "paid_hours",
                                "'Paid hours' is not between 0 and 300", 1))
-  error_rows <- idx[data$paid_hours >= 220 & data$paid_hours <= 300]
+  error_rows <- as.numeric(idx[data$paid_hours >= 220 & data$paid_hours <= 300])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$paid_hours[error_rows], "paid_hours",
                                "'Paid hours' is more than 220", 2))
+
+  # Check for activity_rate and paid_hours to be both non-zero
+  error_rows <- as.numeric(idx[data$activity_rate > 0 & data$paid_hours > 0])
+  errors <- rbind(errors,
+                  build_errors(error_rows, data$personal_number[error_rows],
+                               data$paid_hours[error_rows], "paid_hours",
+                               "'Activity rate' and 'paid hours' are both non-zero", 1))
+
   # Check for non-positive basic wage
   error_rows <- as.numeric(idx[data$basic_wage < 0])
   errors <- rbind(errors,
@@ -124,12 +155,14 @@ check_data <- function(data) {
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$basic_wage[error_rows], "basic_wage",
                                "'Basic wage' is zero", 1))
+
   # Check for negative allowances
   error_rows <- as.numeric(idx[data$allowances < 0])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$allowances[error_rows], "allowances",
                                "'Allowances' are negative", 1))
+
   # Check for negative 13th monthly wage
   error_rows <- as.numeric(idx[data$monthly_wage_13 < 0])
   errors <- rbind(errors,
@@ -137,26 +170,25 @@ check_data <- function(data) {
                                data$monthly_wage_13[error_rows],
                                "monthly_wage_13",
                                "'13th monthly wage' is negative", 1))
-  # Tolerance level for rounding errors
-  tol <- 0.1
+
   # Check for 13th monthly wage exceeding 25% of the basic wage
-  error_rows <- as.numeric(idx[4 * data$monthly_wage_13 -
-                                 data$basic_wage > tol])
+  error_rows <- as.numeric(idx[data$monthly_wage_13 > 0.25*data$basic_wage])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$monthly_wage_13[error_rows],
                                "monthly_wage_13",
                                paste0("'13th monthly wage' exceeds 25% of the ",
-                                      "'basic wage'"), 1))
-  # Check for 13th monthly wage less than 8.5% (1/12) of the basic wage
-  error_rows <- as.numeric(idx[data$basic_wage -
-                                 12 * data$monthly_wage_13 > tol])
+                                      "'basic wage'"), 2))
+
+  # Check for 13th monthly wage undercutting 8% of the basic wage
+  error_rows <- as.numeric(idx[data$monthly_wage_13 < 0.08*data$basic_wage])
   errors <- rbind(errors,
                   build_errors(error_rows, data$personal_number[error_rows],
                                data$monthly_wage_13[error_rows],
                                "monthly_wage_13",
-                               paste0("'13th monthly wage' is less than 8.5% ",
-                                      "(1/12) of the 'basic wage'"), 2))
+                               paste0("'13th monthly wage' is less than 8% of the ",
+                                      "'basic wage'"), 2))
+
   # Check for negative special payments
   error_rows <- as.numeric(idx[data$special_payments < 0])
   errors <- rbind(errors,
@@ -164,6 +196,7 @@ check_data <- function(data) {
                                data$special_payments[error_rows],
                                "special_payments",
                                "'Special payments' are negative", 1))
+
   # Check for weekly hours range (>= 1, <= 100)
   error_rows <- as.numeric(idx[data$weekly_hours < 1 | data$weekly_hours > 100])
   errors <- rbind(errors,
@@ -171,6 +204,7 @@ check_data <- function(data) {
                                data$weekly_hours[error_rows],
                                "weekly_hours",
                                "'Weekly hours' are not between 1 and 100", 1))
+
   # Check for plausible weekly hours range (>= 1, <= 50)
   error_rows <- as.numeric(idx[data$weekly_hours > 50])
   errors <- rbind(errors,
@@ -178,6 +212,7 @@ check_data <- function(data) {
                                data$weekly_hours[error_rows],
                                "weekly_hours", "'Weekly hours' are above 50",
                                2))
+
   # Check for wrong population values
   error_rows <- as.numeric(idx[!(data$population %in% 1:5)])
   errors <- rbind(errors,
@@ -185,7 +220,51 @@ check_data <- function(data) {
                                data$population[error_rows],
                                "population",
                                "'Population' is not between 1 and 5", 1))
-  errors
+
+  # Check for combinations of level_of_requirements and professional_position
+  error_rows <- as.numeric(idx[data$level_of_requirements %in% 6:8 &
+                                 data$professional_position %in% 1:2])
+  errors <- rbind(errors,
+                  build_errors(error_rows, data$personal_number[error_rows],
+                               data$level_of_requirements[error_rows],
+                               "level_of_requirements",
+                               paste("Unusual combination: high professional",
+                                      "position and low level of requirements"), 2))
+
+  # Check for combinations of level_of_requirements and training
+  error_rows <- as.numeric(idx[data$level_of_requirements %in% 1:3 &
+                                 data$training %in% 7:8])
+  errors <- rbind(errors,
+                  build_errors(error_rows, data$personal_number[error_rows],
+                               data$level_of_requirements[error_rows],
+                               "level_of_requirements",
+                               paste("Unusual combination: high level of",
+                                     "requirements and low training"), 2))
+
+  # Check for incorrect level or requirements
+  error_rows <- as.numeric(idx[!(data$level_of_requirements %in% 1:8)])
+  errors <- rbind(errors,
+                  build_errors(error_rows, data$personal_number[error_rows],
+                               data$level_of_requirements[error_rows],
+                               "level_of_requirements",
+                               paste("Level of requirements is not an integer",
+                                     "between 1 and 8"), 1))
+
+  # Check for incorrect professional position
+  error_rows <- as.numeric(idx[!(data$professional_position %in% 1:5)])
+  errors <- rbind(errors,
+                  build_errors(error_rows, data$personal_number[error_rows],
+                               data$professional_position[error_rows],
+                               "professional_position",
+                               paste("Professional position is not an integer",
+                                     "between 1 and 5"), 1))
+
+  if(nrow(errors) > 0){
+    return(errors[order(errors$row), ])
+  }
+  if(nrow(errors) == 0){
+    return(errors)
+  }
 }
 
 #' Builds a dataframe of errors
@@ -212,51 +291,4 @@ build_errors <- function(rows, pers_id, vals, column, description, importance) {
              column = rep(column, n_rows),
              description = rep(description, n_rows),
              importance = rep(importance, n_rows))
-}
-
-#' Cleanup data errors
-#'
-#' \code{cleanup_errors} launches a step-by-step cleanup prompt which goes
-#' over all incorrect rows found in the data (as discovered by
-#' \code{check_data})
-#'
-#' @param data a dataframe of the original data, i.e. the data to which the
-#' \code{errors} parameter refers to
-#' @param errors a dataframe of errors as produced by \code{check_data} to be
-#' cleaned up
-#' @param ignore_plausibility_check a boolean indicating whether the user wants
-#' to ignore implausible data. If \code{TRUE}, the implausible data is
-#' interpreted as being correct.
-#'
-#' @return a dataframe of the original data with cleaned up errors
-#'
-#' @keywords internal
-cleanup_errors <- function(data, errors, ignore_plausibility_check = FALSE) {
-  if (ignore_plausibility_check) {
-    errors <- errors[errors$importance == 1, ]
-    if (nrow(errors) == 0) {
-      return(data)
-    }
-  }
-  # Begin with the cleanup
-  error_rows <- unique(errors$row)
-  message(length(error_rows), " erroneous rows found in the dataset, ",
-          "beginning clean up...")
-  # Iterate over each incorrect row and then over each incorrect column within
-  # this row
-  for (r in error_rows) {
-    err <- errors[errors$row == r, ]
-    pers_id <- unique(err$pers_id)
-    message(rep("-", 80))
-    message("The row #", r, " (personal ID: ", pers_id, ") has ", nrow(err),
-            " invalid/implausible values.")
-    for (i in 1:dim(err)[1]) {
-      message("Error type: ", err$description[i])
-      message(" ", ifelse(err$importance[i] == 2, "implausible", "    invalid"),
-              " value: ", err$value)
-      message("Enter the correct value: ")
-      data[err$row, err$column] <- readline()
-    }
-  }
-  data
 }