+++ data-raw/example_datalist.R
@@ -1,3 +1,3 @@
-datalist_imprimerie <- read_data("data-raw/datalist_imprimerie.xlsx")
+datalist_example <- read_data("data-raw/Example_data_sheet_M1.xlsx")
 
-usethis::use_data(datalist_imprimerie, overwrite = TRUE)
+usethis::use_data(datalist_example, overwrite = TRUE)