+++ tests/testthat.R
@@ -2,3 +2,4 @@ library(testthat)
 library(logib)
 
 test_check("logib")
+