+++ NEWS.md
@@ -1,3 +1,13 @@
+# logib (development version)
+
+* Datalists and URLs have been updated
+* Tolerance threshold has been renamed to limit value
+* Second significance test has been removed
+* Level of requirements can now be coded with eight levels
+* prompt_data_cleanup has been removed
+* Update datalist column names
+
+
 # logib 0.1.3
 
-* Update link to methodology document.
+* Update link to methodology document